Move types for qr code response
