+++ src/main/deltachat/controller.ts
@@ -1,8 +1,8 @@
-import DeltaChat, { C, DeltaChat as DeltaChatNode } from 'deltachat-node'
+import DeltaChat, { C, DeltaChat as DeltaChatNode, Lot } from 'deltachat-node'
 import { app as rawApp } from 'electron'
 import { EventEmitter } from 'events'
 import { getLogger } from '../../shared/logger'
-import { JsonContact, Credentials, AppState } from '../../shared/shared-types'
+import { JsonContact, Credentials, AppState, QrCodeResponse } from '../../shared/shared-types'
 import { maybeMarkSeen } from '../markseenFix'
 import * as mainWindow from '../windows/main'
 import DCAutocrypt from './autocrypt'
@@ -22,7 +22,7 @@ import Extras from './extras'
 import { EventId2EventName as eventStrings } from 'deltachat-node/dist/constants'
 
 import { VERSION, BUILD_TIMESTAMP } from '../../shared/build-info'
-import { Timespans, DAYS_UNTIL_UPDATE_SUGGESTION } from '../../shared/constants'
+import { Timespans, DAYS_UNTIL_UPDATE_SUGGESTION, QrState } from '../../shared/constants'
 import { LocaleData } from '../../shared/localize'
 import tempy from 'tempy'
 import path from 'path'
@@ -300,10 +300,11 @@ export default class DeltaChatController extends EventEmitter {
       const dc = new DeltaChat()
       this.registerEventHandler(dc)
       await dc.open(tempy.directory())
-      const state = dc.checkQrCode(qrCode)
+      const checkQr = await dc.checkQrCode(qrCode)
       this.unregisterEventHandler(dc)
       dc.close()
-      return state
+
+      return checkQr
     }
     console.log(this._dc)
     return this._dc.checkQrCode(qrCode)
