+++ src/renderer/delta-remote.ts
@@ -11,6 +11,7 @@ import {
   Credentials,
   DeltaChatAccount,
   DesktopSettings,
+  QrCodeResponse,
 } from '../shared/shared-types'
 import { MuteDuration } from '../shared/constants'
 import { LocaleData } from '../shared/localize'
@@ -34,14 +35,7 @@ class DeltaRemote {
   call(
     fnName: 'checkQrCode',
     qrCode: string
-  ): Promise<{
-    state: number
-    text1: string
-    text1Meaning: string
-    text2: string
-    timestamp: number
-    id: number
-  }>
+  ): Promise<QrCodeResponse>
   call(fnName: 'getNetworkStatus'): Promise<[boolean, string]>
   // autocrypt ----------------------------------------------------------
   call(fnName: 'autocrypt.initiateKeyTransfer'): Promise<string>
